Also added the script itself which sources the other scripts
The X-local_system.sh should be symlinked into "/etc/profile.d".

This allows us to maintain the script in this shared repository,
while only the symlink is left to be versioned in each local
machine's /etc directory.

Also not that X-local_system.sh does not source the snippets
directly from the /usr/local/share/xworld subdirectories, but rather
the snippets in "/usr/local/etc/profile.d".

The idea is to keep symlinks to selected snippets in
/usr/local/etc/profile.d which allows sourcing a selected subset
of the scripts on a per-site basis.