Transfer site customization framework
Add a directory "site" which shall be symlinked as /usr/local/etc/site
and used as it was located there by other scripts. Alternatively it may
be created as a real directory, and all the files from the
site-directory shall be added as symlinks there.

This "site" directory contains shell snippets which are useful for
selectively being included into one's ~/.profile initialization script
for login shells.

See the README file in that directory about instructions how to use the
snippets properly.

The "site"-directory has been part of a different version-controlled
branch with UUID i99np90tmmdbkelxz73lskk31 (normally checked out
as /usr/local/etc/xworld) for a long time, but has been moved into
THIS branch because the contents are not really site-specific.

Its name is an attribution to the fact that the choice which snippets to
use or not are site-specific and user-specific decisions; a site may
also add actual site-specific scripts to /usr/local/etc/site which are
not part of THIS branch.